+++ src/main/java/com/codetaylor/mc/artisanworktables/modules/worktables/integration/crafttweaker/CTArtisanRecipe.java
@@ -144,7 +144,7 @@ public class CTArtisanRecipe
         }
 
       } else {
-        itemStacks.set(matrixIndices[i], Util.getContainerItem(matrixHandler.getStackInSlot(matrixIndices[i])));
+        itemStacks.set(matrixIndices[i], Util.decrease(matrixHandler.getStackInSlot(matrixIndices[i]).copy(), 1, true));
       }
     }
 
@@ -323,7 +323,7 @@ public class CTArtisanRecipe
     for (int i = 0; i < ingredients.size(); i++) {
       IArtisanIngredient artisanIngredient = ingredients.get(i);
 
-      if (!artisanIngredient.isEmpty()) {
+      if (!artisanIngredient.isEmpty() && artisanIngredient instanceof CTArtisanIngredient) {
         CTArtisanIngredient ingredient = (CTArtisanIngredient) artisanIngredient;
         String mark = ingredient.getIngredient().getMark();
 
@@ -342,7 +342,7 @@ public class CTArtisanRecipe
     for (int i = 0; i < ingredients.size(); i++) {
       IArtisanIngredient artisanIngredient = ingredients.get(i);
 
-      if (!artisanIngredient.isEmpty()) {
+      if (!artisanIngredient.isEmpty() && artisanIngredient instanceof CTArtisanIngredient) {
         CTArtisanIngredient ctArtisanIngredient = (CTArtisanIngredient) artisanIngredient;
         String mark = ctArtisanIngredient.getIngredient().getMark();