+++ docs/changelog.md
@@ -1,3 +1,11 @@
+**1.23.3**
+
+  * **Fixed:**
+    * Crash when crafting vanilla recipe with recipe function in Artisan's table [(#240)](https://github.com/codetaylor/artisan-worktables/issues/240)
+    * Improper handling of remaining items [(#241)](https://github.com/codetaylor/artisan-worktables/issues/241)
+    
+---
+
 **1.23.2**
 
   * **Fixed:**